separate Logger class
