+++ SSD_Project/main.cpp
@@ -1,6 +1,9 @@
-#include "../Logger/Logger.cpp"
 #include "CommandManager.h"
 #include "Ssd.cpp"
+
+#include "../Logger/TimeManager.cpp"
+#include "../Logger/LogFileManager.cpp"
+#include "../Logger/Logger.cpp"
 
 int main(int argc, char* argv[]) {
     Ssd ssd;
