+++ TestShell_Sample-Test/test.cpp
@@ -10,6 +10,8 @@
 #include "../TestShell/testshell.cpp"
 #include "../TestShell/RealSsdDriver.cpp"
 #include "../TestShell/RealFileIo.cpp"
+#include "../Logger/TimeManager.cpp"
+#include "../Logger/LogFileManager.cpp"
 #include "../Logger/Logger.cpp"
 
 using namespace std;
